refactor(nx-governance): Audit and remove Core, CLI, and TypeScript adapter leakage from nx-governance #2
